improving modal experience
