+++ src/components/NavBar/Navbar.tsx
@@ -1,3 +1,4 @@
+import WalletDropdown from 'components/WalletDropdown'
 import Web3Status from 'components/Web3Status'
 import { NftVariant, useNftFlag } from 'featureFlags/flags/nft'
 import { useWindowSize } from 'hooks/useWindowSize'
@@ -108,6 +109,7 @@ const Navbar = () => {
             <MenuDropdown />
             <ChainSwitcher />
             <Web3Status />
+            <WalletDropdown />
           </Row>
         </Box>
       </Box>
